refactor(auth): 스키마 재정의 및 카프카 연결 추가
