+++ src/plugins/kafka.ts
@@ -4,4 +4,7 @@ import * as process from 'node:process';
 export const kafka = new Kafka({ brokers: [process.env.KAFKA_BROKER] });
 export const producer = kafka.producer();
 
-producer.connect();
+export const connectKafkaProducer = async () => {
+  await producer.connect();
+  console.log('Kafka producer connected');
+};
