+++ src/server-config.ts
@@ -6,6 +6,7 @@ import { FastifyInstance } from 'fastify';
 import { fastifyRedis } from '@fastify/redis';
 import swaggerPlugin from './plugins/swagger/swagger-plugin.js';
 import { Server } from 'socket.io';
+import { connectKafkaProducer } from './plugins/kafka.js';
 
 export async function configureServer(server: FastifyInstance) {
   server.setValidatorCompiler(validatorCompiler); // Fastify 유효성 검사기 설정
@@ -17,6 +18,7 @@ export async function registerPlugins(server: FastifyInstance) {
   await registerRedisPlugin(server); // Redis 플러그인 등록
   await setDiContainer(server); // 의존성 주입 컨테이너 설정
   await registerSwaggerPlugin(server); // Swagger 플러그인 등록
+  await connectKafkaProducer(); // Kafka 프로듀서 연결
   await server.register(app, { prefix: '/api' }); // REST API 라우트 등록
 }
