+++ src/v1/apis/auth/auth.controller.ts
@@ -1,26 +1,23 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 
-import {
-  loginRequestSchema,
-  loginResponseSchema,
-  requestVerificationCodeInputSchema,
-  signupRequestSchema,
-} from './auth.schema.js';
 import AuthService from './auth.service.js';
 import { TypeOf } from 'zod';
 import { STATUS } from '../../common/constants/status.js';
+import { signupInputSchema } from './schemas/signup.schema.js';
+import { loginInputSchema, loginResponseSchema } from './schemas/login.schema.js';
+import { requestVerificationCodeInputSchema } from './schemas/requestVerificationCode.schema.js';
 
 export default class AuthController {
   constructor(private readonly authService: AuthService) {}
 
   signup = async (request: FastifyRequest, reply: FastifyReply) => {
-    const body = signupRequestSchema.parse(request.body);
+    const body = signupInputSchema.parse(request.body);
     const result = await this.authService.signup(body);
     reply.status(201).send(result);
   };
 
   login = async (request: FastifyRequest, reply: FastifyReply) => {
-    const body = loginRequestSchema.parse(request.body);
+    const body = loginInputSchema.parse(request.body);
     const { refreshToken, accessToken, refreshTokenExpiresAt } = await this.authService.login(body);
 
     reply.header(
